[SP-121] refactor: directory meeting -> team으로 이동
