+++ src/main/java/com/cupid/jikting/team/controller/TeamController.java
@@ -1,10 +1,11 @@
-package com.cupid.jikting.meeting.controller;
+package com.cupid.jikting.team.controller;
 
-import com.cupid.jikting.meeting.dto.TeamProfileResponse;
-import com.cupid.jikting.meeting.service.MeetingService;
+import com.cupid.jikting.team.dto.TeamProfileResponse;
+import com.cupid.jikting.team.service.TeamService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -12,10 +13,10 @@ import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/meetings")
-public class MeetingController {
+@RequestMapping("/teams")
+public class TeamController {
 
-    private final MeetingService meetingService;
+    private final TeamService teamService;
 
     @GetMapping("/likes/received")
     public ResponseEntity<List<TeamProfileResponse>> getReceivedLikes() {
